done: erweierung des wartebereiches.
todo: verifysfm, check vehicleforce, add pedestirn check for matching timme

TODO
make gui runnable
 add Crowd influence

+++ roundabout-core/src/main/java/at/fhv/itm3/s2/roundabout/entity/StreetSection.java
@@ -749,7 +749,7 @@ public class StreetSection extends Street {
      *
      * @return int.
      */
-    public int getPedestrianCrossingEntryHigh(){
+    public double getPedestrianCrossingEntryHigh(){
         return pedestrianCrossing.getHighOfEntry();
     }
